Revert "Split Editor into seprate files by concern"
This reverts commit b26611ced08c26c8eab1fe31a9104e95e4266297.
