--- src/pages/Editor/styles.scss
@@ -1,296 +1,0 @@
-$brown: #7E7261;
-$dark-brown: #4D3D36;
-$green: #5DBF98;
-$off-white: #F8F3E8;
-$orange: #DC8D69;
-$pink: #F4BFC6;
-$teal: #57B7A8;
-$white: #FFFFFF;
-$yellow: #F8CC61;
-$modal-info: rgba(21, 50, 69, 0.7);
-$modal-normal: rgba(47, 31, 14, 0.9);
-
-button, input[type="button"] {
-  background-color: $brown;
-  border: none;
-  border-radius: 35px;
-  box-shadow: rgba(0,0,0,0.2) 0 0 8px;
-  color: $off-white;
-  cursor: pointer;
-  display: inline-flex;
-  align-items: center;
-  font-size: 13px;
-  font-weight: 800;
-  text-transform: uppercase;
-  padding: 10px 14px;
-
-  &#download-acnl {
-    background-color: $teal;
-    color: $white;
-  }
-
-  &#edit-info-button {
-    background-color: $green;
-    color: $white;
-    margin: 0 0 5px;
-  }
-
-  &#scan-button {
-    background-color: $off-white;
-    color: $brown;
-    // width: 170px;
-  }
-
-  &.drawing-icon {
-    background-color: transparent;
-    border-radius: none;
-    box-shadow: none;
-    height: 100px;
-    width: 100px;
-  }
-}
-
-.modal {
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  display: flex;
-  margin: 50px auto;
-  max-width: 80%;
-  color: $white;
-}
-
-.modal-centered {
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  display: flex;
-}
-
-.modal-header {
-  background-color: $modal-normal;
-  border-radius: 45px 45px 0 0;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 10px 30px 0;
-  min-width: 300px;
-}
-
-.modal-window {
-  background-color: $modal-normal;
-  border-radius: 35px;
-  padding: 25px;
-  min-width: 600px;
-}
-
-.modal-info {
-  display: flex;
-  justify-content: center;
-  width: 100%;
-  position: fixed;
-  top: 55%;
-  font-size: 32px;
-  text-align: center;
-}
-
-.info-text {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  background-color: $modal-info;
-  border-radius: 35px;
-  padding: 40px;
-  min-width: 800px;
-  min-height: 150px;
-  max-width: 60%;
-  max-height: 60%;
-
-  p {
-    width: 350px;
-    display: inline-block;
-    line-height: 40px;
-  }
-
-  span {
-    color: $teal;
-  }
-}
-
-.topbar-buttons {
-  display: inline-flex;
-  align-items: center;
-  justify-content: space-evenly;
-  height: 62px;
-  .menu-button {
-    background-color: $orange;
-    color: $white;
-  }
-}
-
-main {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-
-  #left, #center, #right {
-    display: flex;
-    flex-direction: column;
-  }
-
-  #left {
-    padding-right: 40px;
-  }
-
-  #center canvas, #left canvas {
-    box-shadow: 0px 12px 12px -3px rgba(0,0,0,0.3);
-  }
-
-  .colorPicker-menu {
-    height: 0;
-    position: fixed;
-    z-index: 1;
-    overflow: hidden;
-    transition: 0.5s;
-    top: 60px;
-    border-radius: 0 0 35px 35px;
-    background-color: $pink;
-    left: 50%;
-    transform: translate(-50%);
-    button {
-      display: block;
-      margin: 0 auto;
-    }
-  }
-}
-
-.bottom-buttons {
-  padding: 20px 10% 0 0;
-  text-align: right;
-}
-
-.svg {
-  padding: 5px;
-  pointer-events: none;
-
-  &.nav {
-    border-radius: 100%;
-    margin-right: 5px;
-    height: 25px;
-    width: 25px;
-  }
-
-  &.toolbar{
-    height: 50px;
-    width: 50px;
-  }
-
-  &.white-circle {
-    background-color: $white;
-  }
-
-  &.brown-circle {
-    background-color: $brown;
-  }
-}
-
-.previews {
-  display: inline-flex;
-  flex-direction: column;
-  justify-content: space-between;
-  padding-top: 62px;
-  height: 512px;
-}
-
-.tools-and-colors {
-  height: 512px;
-  display: inline-flex;
-  flex-direction: row;
-  align-items: center;
-  border-radius: 0 35px 35px 0;
-  background-color: $pink;
-}
-
-.tool-buttons {
-  display: inline-flex;
-  flex-direction: column;
-  align-items: right;
-  button{
-    margin: 5px;
-  }
-}
-
-canvas.fordrawing {
-  background: repeating-linear-gradient(-45deg, #ddd, #ddd 5px, #fff 5px, #fff 10px);
-}
-
-.pattern-info {
-  margin: 10px 0;
-  max-width: 196px;
-  overflow: hidden;
-}
-
-.render-preview {
-  border: 3px solid $brown;
-  width: 196px;
-  height: 300px;
-  border-radius: 5px;
-}
-
-#change-info-modal.modal-window {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-
-  .edit-info {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    min-width: 500px;
-    max-width: 60%;
-
-    span {
-      display: flex;
-      flex-direction: row;
-      justify-content: space-between;
-      width: 60%;
-    }
-  }
-
-  .edit-notice {
-    max-width: 600px;
-    margin: 20px;
-  }
-}
-
-#publish-modal.modal-window{
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-
-  ol {
-    list-style: decimal;
-    margin: 10px 0;
-    padding: 0 0 0 30px;
-  }
-
-  .left, .right {
-    flex: 1 1 0;
-    align-items: center;
-    max-width: 400px;
-  }
-
-  .dropdowns {
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-  }
-
-  .dropdown, span {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-  }
-}
