+++ src/routers/offline.js
@@ -1,6 +1,6 @@
 import Vue from "vue";
 import VueRouter from "vue-router";
-import Editor from '/pages/Editor/Editor.vue';
+import Editor from '/pages/Editor.vue';
 Vue.use(VueRouter);
 
 const mode = "history";
